move `printReport` to its own file
