+++ source-code/core/src/lint/context.test.ts
@@ -155,41 +155,3 @@ describe("createContext", async () => {
 		})
 	})
 })
-
-describe("printReport", async () => {
-	test("should not print anything if no report get's passed", async () => {
-
-	})
-
-	describe("information should include", async () => {
-		test("level", async () => {
-
-		})
-
-		test("id", async () => {
-
-		})
-
-		test("message", async () => {
-
-		})
-
-		test("no metadata if not present", async () => {
-
-		})
-
-		test("metadata if present", async () => {
-
-		})
-	})
-
-	describe("level", async () => {
-		test("should use `console.error` on 'error'", async () => {
-
-		})
-
-		test("should use `console.warn` on 'warn'", async () => {
-
-		})
-	})
-})
