+++ source-code/core/src/lint/output.test.ts
@@ -1,0 +1,39 @@
+import { describe, test } from "vitest";
+
+describe("printReport", async () => {
+	test("should not print anything if no report get's passed", async () => {
+
+	})
+
+	describe("information should include", async () => {
+		test("level", async () => {
+
+		})
+
+		test("id", async () => {
+
+		})
+
+		test("message", async () => {
+
+		})
+
+		test("no metadata if not present", async () => {
+
+		})
+
+		test("metadata if present", async () => {
+
+		})
+	})
+
+	describe("level", async () => {
+		test("should use `console.error` on 'error'", async () => {
+
+		})
+
+		test("should use `console.warn` on 'warn'", async () => {
+
+		})
+	})
+})
